Added feature to include the road type in the MATSim output.
This is needed so that MATSim scenarios built on OSM networks can be used with the emissions module, which requires road type.

The format out the output is type_speed(kmh) (i.e. residential_30)

An option has been added to the preferences (Convertor tab) to turn this feature on or off.

The validation of the exported network has been extended to warn of links with no road type.